Simpele Controller structuur (oude Program.cs is nu ProgramOld.cs)
